--- examples/clouddrun-to-bq/requirements-test.txt
@@ -1,2 +1,0 @@
-pytest==7.1.1
-pylint
