+++ examples/cloudrun-to-bq/requirements.txt
@@ -1,0 +1,17 @@
+# Copyright 2022 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+
+Flask==2.1.0
+gunicorn==20.1.0
+google-cloud-bigquery==1.8.1
+google-cloud-core==0.29.1
+google-cloud-logging
+jsonschema
